+++ appseed/c/bsdiff/bzlib.cpp
@@ -998,7 +998,7 @@ void BZ_API(BZ2_bzWrite)
 
       if (bzf->strm.avail_out < BZ_MAX_UNUSED) {
          n = BZ_MAX_UNUSED - bzf->strm.avail_out;
-         n2 = fwrite_dup ( (void *)(bzf->buf), sizeof(UChar), 
+         n2 = (int32_t) fwrite_dup ( (void *)(bzf->buf), sizeof(UChar), 
                        n, bzf->handle );
          if (n != n2 || ferror_dup(bzf->handle))
             { BZ_SETERR(BZ_IO_ERROR); return; };
@@ -1057,7 +1057,7 @@ void BZ_API(BZ2_bzWriteClose64)
 
          if (bzf->strm.avail_out < BZ_MAX_UNUSED) {
             n = BZ_MAX_UNUSED - bzf->strm.avail_out;
-            n2 = fwrite_dup ( (void *)(bzf->buf), sizeof(UChar), 
+            n2 = (int32_t) fwrite_dup ( (void *)(bzf->buf), sizeof(UChar), 
                           n, bzf->handle );
             if (n != n2 || ferror_dup(bzf->handle))
                { BZ_SETERR(BZ_IO_ERROR); return; };
@@ -1192,7 +1192,7 @@ int BZ_API(BZ2_bzRead)
          { BZ_SETERR(BZ_IO_ERROR); return 0; };
 
       if (bzf->strm.avail_in == 0 && !myfeof(bzf->handle)) {
-         n = fread_dup ( bzf->buf, sizeof(UChar), 
+         n = (int32_t) fread_dup ( bzf->buf, sizeof(UChar), 
                      BZ_MAX_UNUSED, bzf->handle );
          if (ferror_dup(bzf->handle))
             { BZ_SETERR(BZ_IO_ERROR); return 0; };
